Move conditioning class to backend
